Show user notifications with a system tray popup
The Process Dashboard has traditionally displayed user notifications
with a window. To avoid stealing the user's keyboard focus, this window
is only displayed when the dashboard is the active window.

With this change, notifications will be displayed in a system tray popup
if the dashboard window is not active. This allows the user to see
notifications in a timely manner, even if they leave the dashboard
hidden for long periods of time (as might be the case when using a
toolbar integrated into an IDE).