Bugfix: avoid zero-length task IDs
A bug in the WBS sync logic was causing some parent components to
receive a two-entry task ID list whose first element was the empty
string. That, in turn, was confusing the merging logic.

This change fixes the root problem in the sync logic, and also guards
against this data pattern in the merge logic.