Accelerate background sync if a foreground sync is requested
When you open the WBS Editor from a personal dashboard and save changes,
a sync operation is kicked off in the background. The background sync is
run with a low priority to avoid triggering excessive CPU utilization.

If the user then explicitly requests a sync while that background sync
is still running, this change will arrange for the background sync to
switch back to high priority.