Revert "Avoid StackOverflowError in DashboardSecurityManager..."
This reverts commit d65ec2f6cdc25ae90fe29033c79fec885c56e283. That
commit was an attempt to avoid triggering a StackOverflowError, but the
change ultimately had no effect. This is because the overflow was not
caused by dashboard logic after all, but by a bug in Java 7 and Java 8:
    https://bugs.java.com/view_bug.do?bug_id=8166366

It does not appear to be possible to work around that Java bug with
application logic; so this change has been reverted. The underlying Java
bug was fixed in Java 9, and will be fixed in Java 8u172.