Bugfix: set effective EV period correctly for automatic schedules
The EV calculation logic will automatically extend the schedule as
needed to produce necessary available hours, or to record actual data.
When the current date/time is within the auto-extended part, and time
had been logged but no tasks completed recently, the "effectivePeriod"
index was being calculated incorrectly. This was causing many charts to
truncate the "Actual" lines.