Use the PD version number to identify the client bridge logic
The bridged client logic passes a version number to the server to
identify itself. In the past, it always sent the hardcoded value "1.0"
(with the intent to signify a communications pattern that was
backwards-compatible to the original DataBridge wire protocol).

That wasn't completely accurate (since new capabilities have been added
over time in a backwards-compatible manner), and it wasn't particularly
informative either. With this change, the Process Dashboard version
number is now sent instead. This will allow better auditing and more
fine-grained decision-making on the server-side.