Update the JNLP relaunching logic to use our internal JNLP support
After an individual has migrated their personal dataset to the PDES,
they might still use their old/legacy shortcut to launch the dashboard.
When this occurs, we relaunch the app using JNLP, so they benefit from
launch profile updates. In the past, this was done using the javaws
application; but since that is going away we use our internal JNLP
launching logic instead.