Refactor authentication logic to survive in-process classloader changes
The JNLP handling logic now performs an in-process relaunch of the
application, creating new classloaders and launching a new main() class
for a different codebase. When this occurs, we don't want to lose any
authentication that the user has already performed by entering
credentials. This change arranges for the HttpAuthenticator to store and
retrieve state via system properties, in the event the HttpAuthenticator
class is reloaded via a new classloader. It also makes certain the
CookieManager is not replaced unnecessarily.