Fork when launching JNLP files on Mac OS X
Mac OS X automatically enforces single-instance mode for applications.
So if an app has been launched (by clicking in the Applications folder,
or otherwise), invoking it again will not launch a second copy. This
conflicts with the process-per-dataset architecture for dashboard
launching, so this change modifies the JNLP launcher to fork a new
process when opening a dataset.