Make use of the AppData\Local directory on Windows
On the Windows operating system, the AppData directory includes a
"Local" subdirectory (for files that cached on the local computer) and a
"Roaming" subdirectory (for user-specific files that should follow the
user when they log in from various computers). This change recognizes
that distinction, and uses "Local" for cached launch profiles and cached
import directories.