Bugfix: Java 11 ignores modifications to user.dir system property
The configureWorkingDirectory method has historically updated the
current working directory so it points to the directory containing data
files. However, Java 11 no loner allows a process to modify the
"user.dir" system property. This change fixes places in the code that
were relying on the value of the current working directory.