Improve handling of logic that bootstraps new dashboard instances
1) When bootstrap-starting a new dashboard instance within the current
process, create a new thread to run the logic on. This allows the new
instance to run with a cleaner stack trace, potentially enabling class
definitions in the current classloader to be garbage collected.

2) If multiple dashboard instances are bootstrapped from a single
launcher (for example, because multiple arguments were passed on the
command line), open the second and subsequent datasets in newly forked
subprocesses.