Clean up handling of renderers in the workflow editor
The workflow editor uses a background color to indicate read-only cells.
In the past, this was handled by installing custom renderers on every
workflow column.  A better approach is to set the background color by
overriding the JTable.prepareRenderer() method, since this allows
background colors to be a table behavior rather than a cell behavior.