Register pdlnk and tdlnk file types when installing on Windows
When a user installs the regular, publicly-available Process Dashboard
distribution on Windows, go ahead and set up the file type associations
for .pdlnk and .tdlnk files.  This will position users to be ready for
the PDES if their organization adopts it at some later time.

(For now, no file type associations are made for Mac or Unix users.
Those users will need to run the PDES launcher installer manually.)