Move screenshots of PSP scripts/forms into the main distribution
In the past, when help topics included screenshots of PSP scripts and
forms, those images had to be placed in the PSP add-on (to comply with
copyright/license constraints). Now that the PSP is freely available
under Creative Commons, that separation is no longer necessary. This
change moves the images back into the main dashboard distribution.