Bugfix: eagerly assign datasetIDs to avoid anonymous dataset prefs
The Settings object maintains a Preferences node for dataset-specific
user prefs.  In the past, datasets that did not have a datasetID.dat
file would use the token "anonymous" as part of this node name.
Unfortunately, this led to a number of problems:
 * all anonymous team and personal dashboards shared the same node,
   resulting in pref collisions (like window size)
 * performing an action that finally assigned the dataset ID would cause
   all prefs to be lost after the next restart.
This change corrects those problems by eagerly assigning dataset IDs for
all new dashboards.