Dynamically select size units for generic components
In the past, the generic "Component" type was just a software component
by another name. As such, the size units column would display "LOC" for
these components. This was confusing to non-software teams.

In the new size data model, people can enter any sizes they want on a
Component. If they enter exactly one size estimate on a component,
automatically adopt that as the size units for the component. If no
units have been entered for a Component, display an empty cell.