After reverse syncing size estimates, make tasks non-rate-driven
Some tasks in the WBS can have time estimates that are produced from
size estimates and rates. If a size estimate is changed in a personal
dashboard and reverse-synced to the WBS, we don't want that change to
ripple into the time estimate. This commit clears rates from affected
tasks to avoid that unintended ripple effect.