Enable WBS-managed size data for all new and relaunched team projects
When PD 2.5.3 or higher is used to create a new project in the Team
Dashboard, or to relaunch an existing project, flag the project and its
WBS as using new-style (i.e. WBS-managed) size data.

Because the old software cannot properly handle WBS-managed size data,
this also marks the Team Dashboard as requiring PD 2.5.3 and Team Tools
5.0.0.