Reject Size Inventory API calls for team projects using WBS-managed size
In the past, a REST API has offered users the ability to add rows to the
Size Inventory Form in a team project. But the Size Inventory Form no
longer works in the same way for a project with WBS-managed size data.
Detect this scenario and reject those REST API requests.