Show PROBE scripts that span a project relaunch boundary
When a workflow includes a PROBE task, the personal dashboard shows a
Size Estimating Template and a Project Plan Summary form for that
component. This code change arranges for those items to be consistently
displayed in the script menu, even when the PROBE task and the currently
active task are in different projects (on either side of a relaunch
boundary).