Add logic to relaunch projects with WBS-managed size data
When a project is using WBS-managed size data, we need the relaunch
operation to be a no-op: that is, a master project rollup before and
after the relaunch should show the same summed-up totals for size. This
requires us to move the size values from the new project and delete them
from the old. This change arranges for that to occur.