Bugfix: time log interrupt logic might be skipped on sleep/wake
If a user stops the timer before leaving work, then resumes work on the
same task the next day, we don't want to show a time log entry with
hours of interrupt time. The time logging model has always included a
check for this. However, that check was being performed once-a-minute in
response to a timer event, so it might be skipped if the computer was
placed in a power-saving mode for the entire interrupt period. This
change avoids that problem by repeating the check when the timer is
restarted.