Extend the Team Project Setup Wizard to create personal projects
Team projects provide extensive functionality, such as the WBS Editor
for planning, Common Team Workflows for custom process definition, etc.
This functionality is useful to individual performers too; but in the
past an individual would have to create a dummy team dashboard to make
use of these features. This change introduces the ability to create a
personal project directly from a personal dashboard.