Display the more sophisticated 'team' plan summary for personal projects
When an individual joins a team project, their personal dashboard only
contains a small slice of the overall team data. In that scenario, a lot
of analyses don't make sense (quality in particular). For that reason,
individuals are shown a streamlined project plan summary.

Those problems don't apply for a personal project. In that scenario, it
is useful to show the full plan summary report, complete with the
Quality Summary and Analysis Charts. This change arranges for the script
menu to display the correct report, based on whether the project is a
personal or team project.