Don't confirm deletions/completions when syncing a personal project
The "Sync to WBS" operation finds edits that have been made to the WBS,
and applies the same changes to the personal plan.  On a team project,
this operation prompts for user confirmation before making destructive
changes; the rationale being that another person might have made those
changes in the WBS, and the individual should have the final say about
their personal plan. Of course, this confirmation step is unnecessary on
a personal project, since the person undoubtedly made the WBS deletions
themselves. This change arranges for the sync operation to proceed
without secondary confirmation within a personal project.