Auto-assign WBS items to the owner of a personal project
In a personal project, all items should implicitly be assigned to the
owner of the project, who happens to be its sole team member. The
"Optimize Editing for..." option would do this partially, but not fully
enough (for example, it wouldn't auto-assign leaf components). This
change introduces a secondary auto-assignment step that is more robust
and aggressive. (This is critically important, because we don't plan to
display the "Assigned To" column, so users would be unable to see or
repair unassigned items.)