Tailor the behavior of the WBS balancing bar for personal projects
On a team project, the balancing bars watch for mouseover events and
automatically highlight a given row and/or milestone. Those highlights
aren't as useful in a personal project, so it's distracting to perform
them automatically on mouseover events.

In addition, the balancing bar tooltips do not need to display the name
of the team member within a personal project.