Avoid/repair data mismatches when Replacing Data into a personal project
If a user is working on a personal project and they use the "File >
Replace Data" operation, they could select a replacement data source
that came from a team project with multiple individuals. Check for that
scenario, and alter incoming data to avoid creating a plan that violates
our personal project data assumptions.