Bugfix: auto-zero users were breaking rate-driven workflows
When you apply a rate-driven workflow to a component with an existing
size estimate, it is supposed to compute times for each task based on
the productivity rate. Unfortunately, this functionality was breaking if
the component had been assigned to an individual before applying the
workflow. In that scenario, the "time per person" column was displaying
the rate-driven values, but the "team time" column was displaying 0.

This was occurring due to an unfortunate interaction between the
rate-driven workflow logic and the auto-zero-user logic. This change
fixes the bug.