Skip "Replace Data" confirmation prompts in an empty unedited project
If the user opens the completely empty WBS in a brand new project and
immediately requests a "Replace Data" operation, skip the scary
confirmation prompts about overwriting data.

This will provide better usability for scenarios when a user wants to
initialize a project from a saved template (for example, at the start of
a training course).