Support personal projects in the PDES auto-import logic
When a current dashboard user opens their PDES-based dataset for the
first time, the dashboard will offer to migrate their existing data into
the server. This change extends that logic to support the migration of
any personal projects they may have created in the past. It relies upon
new logic in the tsimport plugin, which will be bundled into launch
profiles going forward.