Bugfix for the (unpublished) MCF Editor action
For years, the dashboard has included an unpublished action that was
capable of displaying the MCF Editor. Unfortunately, the logic in that
action was broken by the move to the Jetty web server. This change fixes
the broken logic, making it possible for a simple *-template.xml file to
expose the action on the Tools menu.