Allow MCFs to define size metrics without associated work product types
Older versions of the WBS sync logic required a single size value to be
entered on any given WBS component. To enforce this, it created a work
product node type for each metric, with a one-to-one mapping. The work
product chosen for a given WBS node would control which size metric it
used.

Since 2.5.3, size metrics can be entered anywhere in the WBS, and
multiple size values can be attached to a single node. As a result, the
need for distinct work product types has lessened. This change allows an
MCF to define size metrics that have no associated work product type.