Improve the usability of newly created workflow models
When a new project is created, the WBS Editor creates an empty workflow
model. That model contains a single row, representing the header for a
workflow with no tasks inside. In the past, that single row was of type
"Software Component." At face value, that type was irrelevant because
the row gets displayed as a "Workflow" node type instead.

But when the user presses <Enter> to create a new row and <Tab> to
indent, they would see "Software Component" as the suggested type of
this first item in their workflow. That gives the user the wrong idea,
suggesting that it's common for workflows to contain components (when
they virtually always contain tasks).

This change assigns a task type instead of "Software Component" - so
when the user adds the first item to a workflow, it will be a task type.
This subtle change helps guide the user to understand how workflows
should be built.