Enhance the LOC diff report to offer expanding/collapsing regions
The LOC diff report generates an HTML page with redlines. In the past,
when diffs were separated by large sections of base code, it could be
difficult to scroll through the report and find the actual diffs. This
change makes the redlines look more like traditional diff output: each
change is surrounded by a few lines of context, and the intervening
blocks are collapsed. Users can click to expand and collapse these
regions.