When adding a task after a PROBE task, choose a correct phase type
Two PROBE tasks cannot follow each other; so when the user presses Enter
on a PROBE task, we must change the type of the newly inserted node. In
the past, we changed it to "Planning" - but now we cannot be certain
that phase will exist in a non-PSP MCF. Use an appropriate default
instead.