Retain unrecognized attributes in the project settings.xml file
If we republish an existing settings.xml file for a project, make sure
to retain any attributes it contained on the root tag that we don't
recognize. This offers a bit of future-proofing, in case an older
version of the dashboard modifies a file that was written by a newer
version.