Add a parameter for additional permitted users within PDES projects
Introduce a team project parameter that can be used to grant read/write
permission on the team data directory to extra users. Very few teams
should need to use this, but it can provide a workaround if the PDES is
overzealous in its enforcement of permissions for a particular project.