Don't reject PDES URLs as "invalid" when the current user lacks perms
The TeamServerSelector class was originally created to see if a team
server was reachable, and if it was accepting new sessions. In the
initial PDES implementation, any HTTP error meant the server was
unavailable.

In the years since, the PDES will return 403 (Forbidden) errors to
indicate that an individual does not have permission to access a given
collection. But that lack of permission should not be interpreted the
same as a "server unavailable" condition. This change updates the
TeamServerSelector so it understands this distinction.