Remove the MCF name from script menu entries, add to parameters instead
In upcoming releases of the dashboard, the MCF will become a much less
central concept. In anticipation of that, this change removes the MCF
name from the items that appear in the script menu. Since it is still
important to know the MCF for support and troubleshooting purposes, it
is now displayed on the Project Parameters and Settings page.