Bugfix: ext node deletion logic wasn't parent-child aware
If an external node in the WBS is deleted from the external system, we
can delete it from the WBS too - but only if it hasn't been used by the
project. This determination is based on whether the node is still
underneath the "incoming" folder, and if it doesn't have any "plain"
children. Unfortunately, the original implementation of this logic was
not updated when external nodes became hierarchical. This change fixes
that oversight.