Improve traceability when nodes dis/re-appear in our ext node query
When a node disappears from the external system, we delete all related
attributes from the WBS to make it "plain." This is appropriate if the
external node was actually deleted. But if the disappearance stemmed
from the ext node not matching a query "WHERE" clause, it's possible the
node could reappear in the future. To handle this, retain the external
ID of "plainified" nodes, and let the sync logic "reclaim" these nodes
when they reappear in the query results.