When editing a WBS ZIP file, offer "Save As" instead of "Save a Copy"
The "File > Save a Copy" option is appropriate when the user is editing
the WBS for a project. That makes it clear we're making a copy of the
real project and saving it into a file. In contrast, when the user is
editing a ZIP file, it's better to offer a "Save As" option, which is a
more familiar user interface paradigm.