Reorganize the WBS Editor toolbar into two rows
The toolbar in the WBS Editor originally contained a small number of
items, so it was tucked into an unused space in the top-left corner of
the window. But with the addition of new functionality, the number of
toolbar items has grown; and as a result the toolbar is almost always
obscured by the data-table-selection tabs. To avoid obscuring toolbar
buttons, this change breaks the toolbar into two rows.