Prevent two daemon processes from simulaneously syncing the same target
If two different processes run simultanously, both synchronizing the
same WBS to the same external source, arrange for the first process to
detect the conflict and exit.

This avoids problems that arise when two daemons are trying to modify
the same data. It also makes it easy to restart a daemon (for example,
to reread new settings) - just launch the process again and the old
process will exit.