Assign size metric unique IDs in a repeatable way
The unique IDs of nodes in the size metric model will be used to store
planned and actual size data in the main WBS. This commit arranges to
use a consistent ID for a given metric if it is deleted and later
readded. Doing so will allow us to resurrect any size data that might
have been entered before the deletion of the metric.