Bugfix: "Save As" was not clearing the global dirty flag
When the WBS Editor is running against a ZIP file, the File menu shows
"Save As" instead of "Save a Copy." Unfortunately, when the user chose
the "Save As" option, the global dirty flag was not being cleared; so
they were being asked if they want to save changes on a subsequent
shutdown. This corrects that problem.