Bugfix: workflow report in personal dashboard assumes construction step
When the Workflow Analysis Report runs in the personal dashboard, it
seeks to exclude components where the current user's involvement was
limited to peer reviews. To do this, it has looked for time logged to a
construction task.

Unfortunately, if a workflow has no construction tasks (for example, if
it's comprised entirely of "Misc" steps), this check would discard all
enactments, resulting in an always-empty personal workflow report. This
change detects that scenario, and skips the exclusion logic.