Bugfix: copy/paste of WBSNodes between windows was sometimes failing
The WBSClipData class provides a mechanism for rich copy-and-paste of
WBS nodes between WBS Editor windows in two different processes.
Unfortunately, this was failing when the WBS nodes contained non-ASCII
data in their names or attributes. The problem was caused by a buggy,
non-charset-aware method in the teamdash.XMLUtils class.