Normalize the horizontal separators on the script menu
When multiple hierarchical levels are all contributing items to the
script menu, the levels are separated by horizontal lines. There is also
a horizontal line immediately preceding the "Add Link" and "More"
options at the bottom. Unfortunately, this final line had a slightly
different appearance than the others. This change fixes that.