When users need Java, send them to a redirect URL on processdash.com
If an individual does not have Java installed on their computer, we
provide a URL where they can download it. In the past, we sent them to
java.com; but that offers a download of the Oracle JRE, which is no
longer free. Instead, send them to a URL at processdash.com. That URL
will be an alias that redirects them to an appropriate download
location, which we can change over time.