Recover when an ext system doesn't record the ID for an exported node
Some external systems can export nodes from the WBS into the external
system. In the process, the sync logic is supposed to record the ID
which the external system assigned to the exported node. If the logic
failed to do so (for example, because it performed a batch update and
couldn't match inputs to outputs), attempt to recover by matching newly
discovered external nodes to pending export WBS nodes.