Refresh ext data after saving a WBS with externally exported nodes
When the user marks a WBS node for export to an external system, then
saves the WBS, perform an ext-refresh as part of the save. This allows
us to immediately pick up any external ID, key, URL, and other metadata
that was assigned to the node by the external system.