Streamline the update logic for BridgedImportSubdirectory
The BridgedImportSubdirectory class handles PDES sync logic for
directories underneath a dashboard dataset, like "./import". In the
past, when code needed to refresh that subdirectory, it was calling
flushData() on the main working directory. That call was overkill, and
resulted in performance problems and/or deadlocks for datasets with a
lot of files. This commit changes the logic so it only updates the
"import" subdir.