Move portions of "save all data" off the Swing event dispatch thread
In the past, the DashController.saveAllData methods would ask the
dashboard to perform a save all operation, and would execute the entire
thing on the Swing event dispatch thread. Unfortunately, this included
potentially time-consuming operations (such as a flush to server), and
the GUI would be frozen during this time. This change arranges for a
smaller footprint on the Swing event dispatch thread.