Use locally cached files to make backups of PDES datasets
The PDES data bridge offers a "backup" action that creates a ZIP of the
files on the server. That action provided performance improvements
during the early versions of the PDES; but in the current version it is
slower than making a ZIP of the locally cached files, and offers no real
benefit. This change elminates the default use of that server-based
action. (The old behavior can be reenabled with a user setting.)