Improve the display of long paths in the time & defect log editors
The Time Log Editor and Defect Log Editor include a column that displays
the path where an entry was logged. Previously, that was displaying a
left-justified full path, which rarely fit in the column. This resulted
in poor usability, as it was impossible to visually scan the table and
determine where entries were logged.

This change rewrites the display so entries are relative to the
currently selected hierarchy node. It also uses smart abbreviation to
display as much as possible of the deepest parts of the path.