Bugfix: reverse sync non-PROBE size on multiply assigned PSP/PROBE tasks
When a PSP task or PROBE task is assigned to multiple people, special
handling kicks in for WBS editability of size metrics, and these
constraints propagate to the handling of reverse sync data.
Unfortunately, the code was aborting reverse sync for all size metrics
of a shared PSP/PROBE task; but it only should have been applying this
constraint to the size metric that PROBE is using.