Bugfix: launcher could not open multiple datasets on Mac OS X
The Process Dashboard Launcher app allows users to browse for and open
server-based datasets. It also registers itself as a handler for the
pdlnk and tdlnk file types, to open datasets after the user has clicked
a PDES link. Unfortunately, in Mac OS X, the app was only allowing a
single process to run at a time. If any dashboard was open, any attempt
to click on the "Launcher" Application icon, or open a second pdlnk
file, resulted in bringing the existing window to the front. This change
corrects that problem.