Preemptively bump version numbers for custom size metrics functionality
The functionality for custom size metrics is a big change with many
implications for forward and backward compatibility.  Version number
comparisons will be needed to navigate these compatibility questions.
With that in mind, this change is proactively setting the new version
numbers we will use for the dashboard and the WBS Editor, so other code
blocks can start implementing the version tests they require.