Include PSP size data in the [Sized_Object_List] for indiv projects
Historically, the [Sized_Object_List] contained a list of entries on the
Size Inventory Form.  When PROBE tasks were introduced, PROBE size data
was added in this list.  This change adds PSP size data to that list as
well.  This will enable a single set of code to simultaneously address
PSP data, PROBE data, and rows on the Size Inventory form for projects
which do not use WBS-managed size.