Skip PDASH export of unneeded metrics for WBS-managed size projects
On a project with WBS-managed size, rollups are computed with data from
the WBS (rather than data from the Size Inventory), along with
reverse-sync data from the userDump.xml file. Accordingly, sums of size
inventory data don't need to be exported to the PDASH files, because
those exported values are never used. This change skips the export of
those values, allowing exports to complete faster.