Properly save/restore sizes of maximized/iconified frames in GuiPrefs
The GuiPrefs object provides functionality to save and restore the last
size and location of a particular GUI window. This change makes the
logic frame-state-aware. Now it can remember whether the window was
maximized, and restore that state; it can remember the non-iconified
size separately and restore that properly; and it won't save spurious
sizes if a window is iconified.