Remove library/process compatibility checks for proxy tables
When importing and exporting from a library, the WBS has traditionally
checked to see if the library used the same MCF as the current project.
That check is no longer needed for proxy estimation tables, since we can
update the customizable size metrics in the current project to match
anything that is imported.