Remove -F arg to work around libexec/java_home bug on Mac OS X Big Sur
The /usr/libexec/java_home app on Mac OS X "Big Sur" shipped with a
number of serious defects. One is that the app behaves incorrectly if it
is given a "-F" (fail-fast) argument. This change removes that argument
to work around the bug. The dashboard works with extremely old versions
of Java, so dropping the fail-fast flag is unlikely to cause problems.