Write sync-related log output into the -sync.pdash file
Gather the log messages that are generated during the external sync
process, and write them into a file within the sync.pdash file. Also,
allow the user to configure the verbosity of this log via a setting in
externals.xml. These capabilities will assist with troubleshooting of
the ext sync process in the future, since the pdash file is included in
data backups of the team dashboard.