Allow the user to sort tasks by user-entered tags
The Task & Schedule Flat View previously displayed a "Sort by Milestone"
button for reordering the task list. This button has been renamed to
"Sort Tasks" and provides access to a list of different sorting options.
Milestone sorting is still offered if applicable, along with an option
to sort by the values entered in the "Sort" column.