Make TeamProjectDataFactory provide a ResourceCollection, not a dir
The ext sync logic has typically worked against files in a physical
directory, which is retrieved from a WorkingDirectory object. But that
isn't ideal for running the ext sync logic within the Enterprise Server,
as it would require us to write project files to the hard drive
temporarily. This commit begins a set of changes to operate against a
ResourceCollection instead of a filesystem directory.