Arrange for WBS Editor task windows to appear on the same screen
When the user opens a task-specific window (such as the workflow editor,
team member list, etc), make that window appear on the same screen as
the WBS Editor by default. Also save and restore locations relative to
the main WBS Editor window. These changes result in better usability in
a multi-screen environment.