Display 100% on top-level workflow nodes when the cell is selected
Double-clicking a workflow percent will normalize the percentages inside
to 100%.  But if the 100% total value disappears, it might make it
harder for a user to guess what just happened. This change arranges for
the 100% total to be displayed when the user selects the cell in
question. When their cursor leaves the cell, the 100% will be hidden
again; this reduces clutter and also serves to help non-normalized
workflows stand out.