Bugfix to the logic for clearing workflow node types
Nodes contain an attribute indicating whether they were created by a
workflow, and/or whether a workflow was applied directly to them. This
change refines the logic for clearing that metadata.

In the past, we would completely clear that attribute whenever the user
changed the type of a node. But while user edits to node type *should*
break the association that this node was created by a workflow, they
should *NOT* destroy the record that a workflow was applied there.

In the past, row-copying operations in the WBS (from Enter/Insert) would
copy the workflow type of the source item to the newly created one. That
logic is appropriate for tasks within a workflow, but we should NOT
claim that a newly inserted single row is a root where a workflow has
been applied.