Try to auto-assign roles when applying workflows to assigned components
It is a best practice to create leaf components representing large
blocks of work, then assign those to individuals for capacity planning.
If a workflow is assigned to one of those capacity-planned components,
and the workflow includes roles in the "Performed By" column, make a
best-effort attempt to map capacity assignments to roles.

To avoid errors, the auto-role-mapping logic only assigns roles that are
"universal." The first universal role in a workflow is one assigned to
every task. The second universal role is one assigned to every task with
at least two people. The third universal role is one assigned to every
task with at least three people, and so on. Non-universal roles are left
unassigned, allowing the user to make the best decision for each.