Bugfix: ad-hoc data elements starting with '[' confused the parser
Background: an individual had a component with a name like:
    [CR 1234] My Component

In the course of regular operations, the dashboard ended up creating a
data element that started with that name. For example:
    [CR 1234] My Component/node==TAG

Unfortunately, when this was written to the ##.dat file, it caused a
startup exception the next time the dashboard was opened. This is
because the initial '[' caused the datafile parser to expect a "new
style" declaration. This change fixes that problem by escaping initial
'[' characters in ##.dat files.