Bugfix: exception using pdash.data.test[] in JSP EL
When writing JSPs, the dashboard provides a pdash.data object that
exposes elements in the data repository. That object was supposed to
offer a test[] method that looks at an element and returns true or
false, based on the SimpleData.test() method. Unfortunately, that idiom
was triggering a ClassNotFoundException because the build was excluding
the necessary .class file from pspdash.jar.