Add a "lenient filtering" mode for FileResourceCollection
The FileResourceCollection receives a filename filter from its strategy
object. In normal operation, it enforces that filter for all I/O
operations: resource names that don't match the filter effectively do
not exist, and can't be read/written/examined. This change introduces a
flag that can be used to relax this. When relaxed, the collection will
still use the filter to produce its list of contained resources; but it
will allow read/write/info queries to succeed even for files that don't
match the filename filter.