Bugfix: avoid duplicate bundle IDs during fast updates
The FileBundleID includes a timestamp to ensure ordering and uniqueness
of bundles. Unfortunately, if a client publishes two bundle updates
within a second of each other, those distinct bundles could end up with
an ID collision. This change introduces a check for that scenario, and
deconflicts by artificially incrementing the timestamp of the later
bundle by one second.