On dashboard startup, save unpublished files from an earlier session
If the dashboard crashes, it misses the opportunity to save files on
exit. Check for this scenario on the next startup, and save any dirty
files found, along with the log file from the crashed session.

Note that if others have changed the same files in the meantime, their
changes will still take precedence. This dirty detection mechanism will
only recover changes to files that no one else has modified in the
meantime.