Delete the obsolete template synchronization logic
In the distant past, individuals might open the Hierarchy Editor in
their personal dashboard and manually build tasks within a team project.
In support of that endeavor, the team dashboard would write a template
file that made it possible to instantiate the steps from a workflow
within the Hierarchy Editor.

That functionality is long-obsolete. It would currently be broken
anyway, because it did not write the proper workflow source ID values
into each node. This change deletes the obsolete logic so these
templates are no longer written by the Team Dashboard.