Streamline the logic used by the ExportJanitor to clean old files
The ExportJanitor deletes files that were written in the past that are
obsolete.  For example, if a user changes their initials, we need to
delete the PDASH file with the old initials.

Unfortunately, it can be tricky to decide whether a historical file is
truly obsolete, or if it was just an alias for a current file. (For
example, if a file was written in the past via UNC, and now through a
drive letter, the two entries represent the same file and the old path
isn't really obsolete.) Past logic would detect these cases of mistaken
identity by checking export timestamps. If a file was rewritten during
the current Export All operation, its path is an alias and not an
obsolete item.

This change simplifies the logic by performing the deletions up front.
That way, we can aggressively delete potentially obsolete entries, on
the knowledge that they will be restored during the current export run.